requirements.txt
Flask==1.1.2
spotipy==2.17.1
python-dotenv==0.15.0
google-cloud-datastore
google-cloud-storage
bcrypt==3.2.0